fix(bedrock): stabilize streaming chunk metadata
