+++ src/core/providers/bedrock/streaming/mod.rs
@@ -2,6 +2,8 @@
 //!
 //! Handles AWS Event Stream parsing and streaming responses
 
+mod event_stream;
+
 use crate::core::providers::bedrock::model_config::{BedrockApiType, BedrockModelFamily};
 use crate::core::providers::unified_provider::ProviderError;
 use crate::core::types::responses::ChatChunk;
@@ -11,33 +13,7 @@ use serde_json::Value;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
-/// AWS Event Stream message
-#[derive(Debug)]
-pub struct EventStreamMessage {
-    pub headers: Vec<EventStreamHeader>,
-    pub payload: Bytes,
-}
-
-/// Event stream header
-#[derive(Debug)]
-pub struct EventStreamHeader {
-    pub name: String,
-    pub value: HeaderValue,
-}
-
-/// Header value types
-#[derive(Debug)]
-pub enum HeaderValue {
-    String(String),
-    ByteArray(Vec<u8>),
-    Boolean(bool),
-    Byte(i8),
-    Short(i16),
-    Integer(i32),
-    Long(i64),
-    UUID(String),
-    Timestamp(i64),
-}
+pub use event_stream::{EventStreamHeader, EventStreamMessage, HeaderValue};
 
 /// Bedrock streaming response
 pub struct BedrockStream {
@@ -45,6 +21,9 @@ pub struct BedrockStream {
     buffer: Vec<u8>,
     model_family: BedrockModelFamily,
     api_type: BedrockApiType,
+    request_model_id: String,
+    request_id: String,
+    created: i64,
 }
 
 impl BedrockStream {
@@ -53,6 +32,8 @@ impl BedrockStream {
         stream: impl Stream<Item = Result<Bytes, reqwest::Error>> + Send + 'static,
         model_family: BedrockModelFamily,
         api_type: BedrockApiType,
+        request_model_id: impl Into<String>,
+        request_id: impl Into<String>,
     ) -> Self {
         let mapped_stream = stream
             .map(|result| result.map_err(|e| ProviderError::network("bedrock", e.to_string())));
@@ -62,159 +43,10 @@ impl BedrockStream {
             buffer: Vec::new(),
             model_family,
             api_type,
+            request_model_id: request_model_id.into(),
+            request_id: request_id.into(),
+            created: chrono::Utc::now().timestamp(),
         }
-    }
-
-    /// Parse event stream message from bytes
-    fn parse_event_message(data: &[u8]) -> Result<EventStreamMessage, ProviderError> {
-        if data.len() < 16 {
-            return Err(ProviderError::response_parsing(
-                "bedrock",
-                "Invalid event stream message",
-            ));
-        }
-
-        // Parse prelude (12 bytes)
-        let total_length = u32::from_be_bytes([data[0], data[1], data[2], data[3]]) as usize;
-        let headers_length = u32::from_be_bytes([data[4], data[5], data[6], data[7]]) as usize;
-        // let prelude_crc = u32::from_be_bytes([data[8], data[9], data[10], data[11]]);
-
-        if data.len() < total_length {
-            return Err(ProviderError::response_parsing(
-                "bedrock",
-                "Incomplete event stream message",
-            ));
-        }
-
-        // Parse headers
-        let mut headers = Vec::new();
-        let mut offset = 12;
-        let headers_end = 12 + headers_length;
-
-        while offset < headers_end {
-            if offset + 1 > data.len() {
-                break;
-            }
-
-            let name_length = data[offset] as usize;
-            offset += 1;
-
-            if offset + name_length > data.len() {
-                break;
-            }
-
-            let name = String::from_utf8_lossy(&data[offset..offset + name_length]).to_string();
-            offset += name_length;
-
-            if offset >= data.len() {
-                break;
-            }
-
-            let header_type = data[offset];
-            offset += 1;
-
-            let value = match header_type {
-                5 | 7 => {
-                    // String type
-                    if offset + 2 > data.len() {
-                        break;
-                    }
-                    let string_length =
-                        u16::from_be_bytes([data[offset], data[offset + 1]]) as usize;
-                    offset += 2;
-                    if offset + string_length > data.len() {
-                        break;
-                    }
-                    let string_value =
-                        String::from_utf8_lossy(&data[offset..offset + string_length]).to_string();
-                    offset += string_length;
-                    HeaderValue::String(string_value)
-                }
-                _ => {
-                    // Skip unknown header types
-                    HeaderValue::String(String::new())
-                }
-            };
-
-            headers.push(EventStreamHeader { name, value });
-        }
-
-        // Extract payload
-        let payload_start = headers_end;
-        let payload_end = total_length - 4; // Exclude message CRC
-        let payload = if payload_start < payload_end && payload_end <= data.len() {
-            Bytes::copy_from_slice(&data[payload_start..payload_end])
-        } else {
-            Bytes::new()
-        };
-
-        Ok(EventStreamMessage { headers, payload })
-    }
-
-    fn header_value<'a>(message: &'a EventStreamMessage, name: &str) -> Option<&'a str> {
-        message.headers.iter().find_map(|header| {
-            (header.name == name)
-                .then_some(&header.value)
-                .and_then(|value| match value {
-                    HeaderValue::String(value) => Some(value.as_str()),
-                    _ => None,
-                })
-        })
-    }
-
-    fn stream_exception_from_payload(value: &Value) -> Option<(String, String)> {
-        let object = value.as_object()?;
-
-        for (code, detail) in object {
-            if code.ends_with("Exception") || code.ends_with("exception") {
-                let message = detail
-                    .get("message")
-                    .and_then(Value::as_str)
-                    .or_else(|| detail.as_str())
-                    .unwrap_or("");
-                return Some((code.clone(), message.to_string()));
-            }
-        }
-
-        None
-    }
-
-    fn stream_error(code: &str, message: &str) -> ProviderError {
-        let details = if message.is_empty() {
-            format!("Bedrock stream error: {code}")
-        } else {
-            format!("Bedrock stream error {code}: {message}")
-        };
-
-        if code.eq_ignore_ascii_case("validationException") {
-            ProviderError::invalid_request("bedrock", details)
-        } else {
-            ProviderError::api_error("bedrock", 500, details)
-        }
-    }
-
-    fn check_stream_error(message: &EventStreamMessage) -> Result<(), ProviderError> {
-        let message_type = Self::header_value(message, ":message-type");
-        let exception_type = Self::header_value(message, ":exception-type");
-
-        if matches!(message_type, Some("exception" | "error")) || exception_type.is_some() {
-            let payload = serde_json::from_slice::<Value>(&message.payload).ok();
-            let payload_message = payload
-                .as_ref()
-                .and_then(|value| value.get("message"))
-                .and_then(Value::as_str)
-                .unwrap_or("");
-            let code = exception_type.unwrap_or("streamException");
-            return Err(Self::stream_error(code, payload_message));
-        }
-
-        if let Ok(payload) = serde_json::from_slice::<Value>(&message.payload)
-            && let Some((code, message)) = Self::stream_exception_from_payload(&payload)
-        {
-            return Err(Self::stream_error(&code, &message));
-        }
-
-        Ok(())
     }
 
     /// Parse chunk based on model family
@@ -366,6 +198,29 @@ impl BedrockStream {
             .unwrap_or(0)
     }
 
+    fn chunk(
+        &self,
+        delta: crate::core::types::responses::ChatDelta,
+        finish_reason: Option<crate::core::types::responses::FinishReason>,
+    ) -> ChatChunk {
+        use crate::core::types::responses::ChatStreamChoice;
+
+        ChatChunk {
+            id: self.request_id.clone(),
+            object: "chat.completion.chunk".to_string(),
+            created: self.created,
+            model: self.request_model_id.clone(),
+            choices: vec![ChatStreamChoice {
+                index: 0,
+                delta,
+                finish_reason,
+                logprobs: None,
+            }],
+            usage: None,
+            system_fingerprint: None,
+        }
+    }
+
     fn parse_converse_tool_start(
         event: &Value,
     ) -> Option<crate::core::types::responses::ToolCallDelta> {
@@ -415,54 +270,34 @@ impl BedrockStream {
 
     /// Parse ConverseStream event chunks.
     fn parse_converse_chunk(&self, value: &Value) -> Result<Option<ChatChunk>, ProviderError> {
-        use crate::core::types::responses::{ChatDelta, ChatStreamChoice};
+        use crate::core::types::responses::ChatDelta;
 
         if let Some(tool_call) = Self::parse_converse_tool_start(value) {
-            return Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
-                        role: None,
-                        content: None,
-                        thinking: None,
-                        tool_calls: Some(vec![tool_call]),
-                        function_call: None,
-                        audio: None,
-                    },
-                    finish_reason: None,
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            }));
+            return Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: None,
+                    thinking: None,
+                    tool_calls: Some(vec![tool_call]),
+                    function_call: None,
+                    audio: None,
+                },
+                None,
+            )));
         }
 
         if let Some(tool_call) = Self::parse_converse_tool_input(value) {
-            return Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
-                        role: None,
-                        content: None,
-                        thinking: None,
-                        tool_calls: Some(vec![tool_call]),
-                        function_call: None,
-                        audio: None,
-                    },
-                    finish_reason: None,
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            }));
+            return Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: None,
+                    thinking: None,
+                    tool_calls: Some(vec![tool_call]),
+                    function_call: None,
+                    audio: None,
+                },
+                None,
+            )));
         }
 
         if let Some(content) = value
@@ -471,27 +306,17 @@ impl BedrockStream {
             .and_then(|d| d.get("text"))
             .and_then(|t| t.as_str())
         {
-            return Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
-                        role: None,
-                        content: Some(content.to_string()),
-                        thinking: None,
-                        tool_calls: None,
-                        function_call: None,
-                        audio: None,
-                    },
-                    finish_reason: None,
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            }));
+            return Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: Some(content.to_string()),
+                    thinking: None,
+                    tool_calls: None,
+                    function_call: None,
+                    audio: None,
+                },
+                None,
+            )));
         }
 
         if let Some(message_stop) = value.get("messageStop") {
@@ -501,27 +326,17 @@ impl BedrockStream {
                 .and_then(|reason| reason.as_str());
             let finish_reason = Self::parse_converse_finish_reason(stop_reason);
 
-            return Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
-                        role: None,
-                        content: None,
-                        thinking: None,
-                        tool_calls: None,
-                        function_call: None,
-                        audio: None,
-                    },
-                    finish_reason: Some(finish_reason),
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            }));
+            return Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: None,
+                    thinking: None,
+                    tool_calls: None,
+                    function_call: None,
+                    audio: None,
+                },
+                Some(finish_reason),
+            )));
         }
 
         Ok(None)
@@ -529,7 +344,7 @@ impl BedrockStream {
 
     /// Parse Claude streaming chunk
     fn parse_claude_chunk(&self, value: &Value) -> Result<Option<ChatChunk>, ProviderError> {
-        use crate::core::types::responses::{ChatDelta, ChatStreamChoice};
+        use crate::core::types::responses::ChatDelta;
 
         // Claude uses specific event types
         let event_type = value.get("type").and_then(|v| v.as_str());
@@ -542,56 +357,36 @@ impl BedrockStream {
                     .and_then(|t| t.as_str())
                     .unwrap_or("");
 
-                Ok(Some(ChatChunk {
-                    id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                    object: "chat.completion.chunk".to_string(),
-                    created: chrono::Utc::now().timestamp(),
-                    model: String::new(),
-                    choices: vec![ChatStreamChoice {
-                        index: 0,
-                        delta: ChatDelta {
-                            role: None,
-                            content: Some(delta.to_string()),
-                            thinking: None,
-                            tool_calls: None,
-                            function_call: None,
-                            audio: None,
-                        },
-                        finish_reason: None,
-                        logprobs: None,
-                    }],
-                    usage: None,
-                    system_fingerprint: None,
-                }))
-            }
-            Some("message_stop") => Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
+                Ok(Some(self.chunk(
+                    ChatDelta {
                         role: None,
-                        content: None,
+                        content: Some(delta.to_string()),
                         thinking: None,
                         tool_calls: None,
                         function_call: None,
                         audio: None,
                     },
-                    finish_reason: Some(crate::core::types::responses::FinishReason::Stop),
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            })),
+                    None,
+                )))
+            }
+            Some("message_stop") => Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: None,
+                    thinking: None,
+                    tool_calls: None,
+                    function_call: None,
+                    audio: None,
+                },
+                Some(crate::core::types::responses::FinishReason::Stop),
+            ))),
             _ => Ok(None),
         }
     }
 
     /// Parse Nova streaming chunk
     fn parse_nova_chunk(&self, value: &Value) -> Result<Option<ChatChunk>, ProviderError> {
-        use crate::core::types::responses::{ChatDelta, ChatStreamChoice};
+        use crate::core::types::responses::ChatDelta;
 
         if let Some(content) = value
             .get("contentBlockDelta")
@@ -599,27 +394,17 @@ impl BedrockStream {
             .and_then(|d| d.get("text"))
             .and_then(|t| t.as_str())
         {
-            Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
-                        role: None,
-                        content: Some(content.to_string()),
-                        thinking: None,
-                        tool_calls: None,
-                        function_call: None,
-                        audio: None,
-                    },
-                    finish_reason: None,
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            }))
+            Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: Some(content.to_string()),
+                    thinking: None,
+                    tool_calls: None,
+                    function_call: None,
+                    audio: None,
+                },
+                None,
+            )))
         } else {
             Ok(None)
         }
@@ -627,34 +412,26 @@ impl BedrockStream {
 
     /// Parse Titan streaming chunk
     fn parse_titan_chunk(&self, value: &Value) -> Result<Option<ChatChunk>, ProviderError> {
-        use crate::core::types::responses::{ChatDelta, ChatStreamChoice};
+        use crate::core::types::responses::ChatDelta;
 
         if let Some(content) = value.get("outputText").and_then(|t| t.as_str()) {
-            Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
-                        role: None,
-                        content: Some(content.to_string()),
-                        thinking: None,
-                        tool_calls: None,
-                        function_call: None,
-                        audio: None,
-                    },
-                    finish_reason: if value.get("completionReason").is_some() {
-                        Some(crate::core::types::responses::FinishReason::Stop)
-                    } else {
-                        None
-                    },
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            }))
+            let finish_reason = if value.get("completionReason").is_some() {
+                Some(crate::core::types::responses::FinishReason::Stop)
+            } else {
+                None
+            };
+
+            Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: Some(content.to_string()),
+                    thinking: None,
+                    tool_calls: None,
+                    function_call: None,
+                    audio: None,
+                },
+                finish_reason,
+            )))
         } else {
             Ok(None)
         }
@@ -662,7 +439,7 @@ impl BedrockStream {
 
     /// Parse generic streaming chunk
     fn parse_generic_chunk(&self, value: &Value) -> Result<Option<ChatChunk>, ProviderError> {
-        use crate::core::types::responses::{ChatDelta, ChatStreamChoice};
+        use crate::core::types::responses::ChatDelta;
 
         let openai_choice = value
             .get("choices")
@@ -711,27 +488,17 @@ impl BedrockStream {
         });
 
         if content.is_some() || openai_tool_calls.is_some() || openai_finish_reason.is_some() {
-            Ok(Some(ChatChunk {
-                id: format!("bedrock-{}", uuid::Uuid::new_v4()),
-                object: "chat.completion.chunk".to_string(),
-                created: chrono::Utc::now().timestamp(),
-                model: String::new(),
-                choices: vec![ChatStreamChoice {
-                    index: 0,
-                    delta: ChatDelta {
-                        role: None,
-                        content: content.map(str::to_string),
-                        thinking: None,
-                        tool_calls: openai_tool_calls,
-                        function_call: None,
-                        audio: None,
-                    },
-                    finish_reason: openai_finish_reason,
-                    logprobs: None,
-                }],
-                usage: None,
-                system_fingerprint: None,
-            }))
+            Ok(Some(self.chunk(
+                ChatDelta {
+                    role: None,
+                    content: content.map(str::to_string),
+                    thinking: None,
+                    tool_calls: openai_tool_calls,
+                    function_call: None,
+                    audio: None,
+                },
+                openai_finish_reason,
+            )))
         } else {
             Ok(None)
         }
